Rename package according to PEP8
Python packages should have short, all-lowercase names,
although the use of underscores is discouraged.
