+++ bundlebuilder/models/bundle.py
