+++ bundlebuilder/models/entity.py
