+++ bundlebuilder/models/indicator.py
