+++ bundlebuilder/version.py
